fix some bug

+++ basemodule/src/main/java/cn/wsgwz/basemodule/utilities/DensityUtil.java
@@ -31,9 +31,9 @@ public class DensityUtil
 	 * @param spVal
 	 * @return
 	 */
-	public static int sp2px( float spVal)
+	public static float sp2px( float spVal)
 	{
-		return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
+		return  TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
 				spVal, Resources.getSystem().getDisplayMetrics());
 	}
